+++ comments/plugin.js
@@ -1,10 +1,9 @@
 import blockComments from "./rules/block-comments.js";
 import classDocumentation from "./rules/class-documentation.js";
 import configuredApiCalls from "./rules/configured-api-calls.js";
+import formatting from "./rules/formatting.js";
 import functionDocumentation from "./rules/function-documentation.js";
 import jsdocTagFormatting from "./rules/jsdoc-tag-formatting.js";
-import lineComments from "./rules/line-comments.js";
-import maxLineLength from "./rules/max-line-length.js";
 import placement from "./rules/placement.js";
 import sentencePunctuation from "./rules/sentence-punctuation.js";
 import variableDeclarations from "./rules/variable-declarations.js";
@@ -18,10 +17,9 @@ export default {
 		"block-comments": blockComments,
 		"class-documentation": classDocumentation,
 		"configured-api-calls": configuredApiCalls,
+		formatting,
 		"function-documentation": functionDocumentation,
 		"jsdoc-tag-formatting": jsdocTagFormatting,
-		"line-comments": lineComments,
-		"max-line-length": maxLineLength,
 		placement,
 		"sentence-punctuation": sentencePunctuation,
 		"variable-declarations": variableDeclarations,