+++ test/comments/oxlint-integration/file-disable.js
@@ -1,6 +1,6 @@
 /* oxlint-disable @stylistic/padding-line-between-statements */
 /* oxlint-disable comments/sentence-punctuation */
-/* oxlint-disable comments/max-line-length */
+/* oxlint-disable comments/formatting */
 /* oxlint-disable comments/placement */
 /* oxlint-disable comments/block-comments */
 /* oxlint-disable comments/function-documentation */