+++ test/comments/placement.test.js
@@ -73,12 +73,12 @@ registerDialog();`,
 		},
 		{
 			name: "does not treat comments separated by a directive as continuations",
-			code: "if (isReady) {\n\t// First comment.\n\t// oxlint-disable-next-line comments/line-comments\n  // Second comment.\n\trunTask();\n}",
+			code: "if (isReady) {\n\t// First comment.\n\t// oxlint-disable-next-line comments/formatting\n  // Second comment.\n\trunTask();\n}",
 			errors: [
 				{ message: "Comment must be immediately before the documented code.", line: 4, column: 2 },
 			],
 			output:
-				"if (isReady) {\n\t// First comment.\n\t// oxlint-disable-next-line comments/line-comments\n\t// Second comment.\n\trunTask();\n}",
+				"if (isReady) {\n\t// First comment.\n\t// oxlint-disable-next-line comments/formatting\n\t// Second comment.\n\trunTask();\n}",
 		},
 		{
 			name: "reindents every line of a wrapped continuation comment to match its declaration",